fix: entity-service 단 변경
